refactor: rename watch... components to movie...
