+++ src/ui/MovieDescription/MovieDescription.module.scss
@@ -6,7 +6,7 @@ $transparentWite: rgba(
   $alpha: 0.72,
 );
 
-.watchDescription {
+.movieDescription {
   line-height: 22px;
   overflow: hidden;
 }
