+++ src/ui/MovieParams/MovieParams.module.scss
@@ -7,7 +7,7 @@ $transparentWhite: rgba(
   $alpha: 0.72,
 );
 
-.watchParams {
+.movieParams {
   display: flex;
   flex-direction: column;
   gap: fn.toRem(4px);
