--- src/ui/WatchParams/index.ts
@@ -1,1 +1,0 @@
-export { WatchParams } from './WatchParams'
